feat: separate 'pull'
